refactored and cleaned code
