+++ src/graphs/general/overallGraphs.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package graphs
+package general
 
 import (
 	"fmt"
@@ -26,98 +26,19 @@ import (
 	"time"
 
 	ui "github.com/gizak/termui/v3"
-	"github.com/gizak/termui/v3/widgets"
 )
 
 var isCPUSet = false
 
 var run = true
 
-type mainPage struct {
-	Grid         *ui.Grid
-	MemoryChart  *widgets.BarChart
-	DiskChart    *widgets.Table
-	NetworkChart *widgets.Plot
-	CPUCharts    []*widgets.Gauge
-	NetPara      *widgets.Paragraph
-}
-
-func newPage(numCores int) *mainPage {
-	page := &mainPage{
-		Grid:         ui.NewGrid(),
-		MemoryChart:  widgets.NewBarChart(),
-		DiskChart:    widgets.NewTable(),
-		NetworkChart: widgets.NewPlot(),
-		CPUCharts:    make([]*widgets.Gauge, 0),
-		NetPara:      widgets.NewParagraph(),
-	}
-	page.init(numCores)
-	return page
-}
-
-func (page *mainPage) init(numCores int) {
-
-	// Initialize Bar Graph for Memory Chart
-	page.MemoryChart.Title = " Memory (RAM) "
-	page.MemoryChart.Labels = []string{"Total", "Available", "Used", "Free"}
-	page.MemoryChart.BarWidth = 8
-	page.MemoryChart.BarGap = 9
-	page.MemoryChart.BarColors = []ui.Color{ui.ColorCyan, ui.ColorGreen}
-	page.MemoryChart.LabelStyles = []ui.Style{ui.NewStyle(ui.ColorWhite)}
-	page.MemoryChart.NumStyles = []ui.Style{ui.NewStyle(ui.ColorBlack)}
-	page.MemoryChart.BorderStyle.Fg = ui.ColorCyan
-
-	// Initialize Table for Disk Chart
-	page.DiskChart.Title = " Disk "
-	page.DiskChart.TextStyle = ui.NewStyle(ui.ColorWhite)
-	page.DiskChart.TextAlignment = ui.AlignLeft
-	page.DiskChart.RowSeparator = false
-	page.DiskChart.ColumnWidths = []int{9, 9, 9, 9, 9, 11}
-	page.DiskChart.BorderStyle.Fg = ui.ColorCyan
-
-	// Initialize Plot for Network Chart
-	page.NetworkChart.Title = " Network data(in mB) "
-	page.NetworkChart.HorizontalScale = 1
-	page.NetworkChart.AxesColor = ui.ColorCyan
-	page.NetworkChart.LineColors[0] = ui.ColorRed
-	page.NetworkChart.LineColors[1] = ui.ColorGreen
-	page.NetworkChart.DrawDirection = widgets.DrawLeft
-	page.NetworkChart.BorderStyle.Fg = ui.ColorCyan
-	page.NetworkChart.DataLabels = []string{"ip kB", "op kB"} //refer issue #214 for details
-
-	// Initialize paragraph for NetPara
-	page.NetPara.Text = "[Total RX](fg:red): 0\n\n[Total TX](fg:green): 0"
-	page.NetPara.Border = true
-	page.NetPara.BorderStyle.Fg = ui.ColorCyan
-	page.NetPara.Title = " RX/TX "
-
-	// Initialize Gauges for each CPU Core usage
-	for i := 0; i < numCores; i++ {
-		tempGauge := widgets.NewGauge()
-		tempGauge.Title = " CPU " + strconv.Itoa(i) + " "
-		tempGauge.Percent = 0
-		tempGauge.BarColor = ui.ColorBlue
-		tempGauge.BorderStyle.Fg = ui.ColorCyan
-		tempGauge.TitleStyle.Fg = ui.ColorWhite
-		page.CPUCharts = append(page.CPUCharts, tempGauge)
-	}
-
-	// Initialize Grid layout
-	page.Grid.Set(
-		ui.NewRow(0.34, page.MemoryChart),
-		ui.NewRow(0.34,
-			ui.NewCol(0.25, page.NetPara),
-			ui.NewCol(0.75, page.NetworkChart),
-		),
-		ui.NewRow(0.34, page.DiskChart),
-	)
-
-	w, h := ui.TerminalDimensions()
-	page.Grid.SetRect(w/2, 0, w, h)
-}
-
 // RenderCharts handles plotting graphs and charts for system stats in general.
-func RenderCharts(endChannel chan os.Signal, memChannel chan []float64, cpuChannel chan []float64, diskChannel chan [][]string, netChannel chan map[string][]float64, wg *sync.WaitGroup) {
+func RenderCharts(endChannel chan os.Signal,
+	memChannel chan []float64,
+	cpuChannel chan []float64,
+	diskChannel chan [][]string,
+	netChannel chan map[string][]float64,
+	wg *sync.WaitGroup) {
 
 	if err := ui.Init(); err != nil {
 		log.Fatalf("failed to initialize termui: %v", err)
@@ -132,7 +53,7 @@ func RenderCharts(endChannel chan os.Signal, memChannel chan []float64, cpuChann
 	isCPUSet = true
 
 	// Create new page
-	myPage := newPage(numCores)
+	myPage := NewPage(numCores)
 
 	// Initialize slices for Network Data
 	ipData := make([]float64, 40)
